mint button refactored to separated component
